pyproject.toml
[project]
name = "product-space"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "anthropic>=0.64.0",
    "black>=25.1.0",
    "crewai>=0.165.1",
    "crewai-tools>=0.62.3",
    "databases[sqlite]>=0.9.0",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.34.4",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.2.14",
    "mypy>=1.17.1",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "regex>=2025.7.34",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "textblob>=0.19.0",
    "uvicorn[standard]>=0.35.0",
    "vadersentiment>=3.3.2",
]
